+++ app/Support/Str/functions/case-converter.php
