+++ app/Support/Str/functions/inflector.php
