+++ inc/functions.php
@@ -2,6 +2,6 @@
 
 declare(strict_types=1);
 
-require_once dirname(__DIR__) . '/app/Support/Str/CaseConverter/functions.php';
-require_once dirname(__DIR__) . '/app/Support/Str/Inflector/functions.php';
+require_once dirname(__DIR__) . '/app/Support/Str/functions/case-converter.php';
+require_once dirname(__DIR__) . '/app/Support/Str/functions/inflector.php';
 require_once dirname(__DIR__) . '/app/Validator/functions.php';
